todo_project/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
drf-rw-serializers==1.1.1
drf-model-serializer==0.0.2
django-hashid-field==3.4.0
django-postgres-extra==2.0.8
django-filter==24.2
django-allauth==0.61.1
requests==2.31.0
psycopg2-binary==2.9.9
python-decouple==3.8
model-bakery==1.18.0
pytest-cov==5.0.0
channels==4.1.0
daphne==4.1.2
